% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{AddPredictions}
\alias{AddPredictions}
\title{Add Predictions}
\usage{
AddPredictions(
  object,
  preds,
  preds.levels = unique(x = as.character(x = preds)),
  preds.drop = TRUE
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{preds}{dataframe: Rows are cells. Columns are predicted.id,
predicted.id.score, and prediction.score. [CLASSNAME] for all reference
classes}

\item{preds.levels}{Levels for predicted IDs, useful for
ordering \code{preds}}

\item{preds.drop}{Drop unused levels from \code{preds} and
\dQuote{predictions} assay.}
}
\value{
\code{object} with predicted.id and predicted.id.score added to
metadata, and prediction.score. [CLASSNAME] added as an assay named
\dQuote{predictions}
}
\description{
Add Predictions
}
\keyword{internal}
