% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{AttachDeps}
\alias{AttachDeps}
\title{Attach dependent packages}
\usage{
AttachDeps()
}
\value{
Attaches the required packages and returns invisible \code{NULL}
}
\description{
Attaches the following packages
\itemize{
 \item shinyBS
}
}
\keyword{internal}
