% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Azimuth-package}
\alias{Azimuth}
\alias{Azimuth-package}
\title{Azimuth: A Shiny App Demonstrating a Query-Reference Mapping Algorithm for Single-Cell Data}
\description{
What the package does (one paragraph).
}
\section{Package options}{


\pkg{Azimuth} uses the following options to control the behavior of the app.
Users can provide these as named arguments to \code{\link{AzimuthApp}}
through dots (...), specify these in the config file, or configure these with
\code{\link[base]{options}}.

\subsection{App options}{
The following options control app behavior
\describe{
\item{\code{Azimuth.app.default_adt}}{
ADT to select by default in feature/violin plot
}
\item{\code{Azimuth.app.default_gene}}{
Gene to select by default in feature/violin plot
}
\item{\code{Azimuth.app.default_metadata}}{
Default metadata transferred from reference.
}
\item{\code{Azimuth.app.demodataset}}{
Path to data file (in any Azimuth-supported format) to automatically load
when the user clicks a button. The button is only available in the UI
if this option is set
}
\item{\code{Azimuth.app.googlesheet}}{
Google Sheet identifier (appropriate for use with
\code{\link[googlesheets4:gs4_get]{googlesheets4:gs4_get}}) to write log
records. Logging is only enabled if this and other \code{google*} options
are set
}
\item{\code{Azimuth.app.googletoken}}{
Path to directory containing Google Authentication token file.
Logging is only enabled if this and other \code{google*} options are set
}
\item{\code{Azimuth.app.googletokenemail}}{
Email address corresponding to the Google Authentication token file.
Logging is only enabled if this and other \code{google*} options are set
}
\item{\code{Azimuth.app.max_cells}}{
Maximum number of cells allowed to upload
}
\item{\code{Azimuth.app.metadata_notransfer}}{
Metadata to annotate in reference but not transfer to query
}
\item{\code{Azimuth.app.mito}}{
Regular expression pattern indicating mitochondrial features in query
object
}
\item{\code{Azimuth.app.plotseed}}{
Seed to shuffle colors for cell types
}
\item{\code{Azimuth.app.reference}}{
URL or directory path to reference dataset; see
\code{\link{LoadReference}} for more details
}
\item{\code{Azimuth.app.refuri}}{
URL for publicly available reference dataset, used in the downloadable
analysis script in case \code{Azimuth.app.reference} points to a directory
}
\item{\code{Azimuth.app.refdescriptor}}{
Provide (as a string) the html to render the reference description on the
welcome page
}
\item{\code{Azimuth.app.welcomebox}}{
Provide (as a string) the code to render the box on the welcome page
(quotes escaped). Example:\preformatted{box(
  h3(\\"Header\\"),
  \\"body text\\",
  a(\\"link\\", href=\\"www.satijalab.org\\", target=\\"_blank\\"),
  width = 12
)
}

}
}
}

\subsection{Control options}{
These options control mapping and analysis behavior
\describe{
\item{\code{Azimuth.map.ncells}}{
Minimum number of cells required to accept uploaded file
defaults to \code{100}
}
\item{\code{Azimuth.map.ngenes}}{
Minimum number of genes in common with reference to accept uploaded file;
defaults to \code{250}
}
\item{\code{Azimuth.map.nanchors}}{
Minimum number of anchors that must be found to complete mapping.
Defaults to \code{50}
}
\item{\code{Azimuth.map.panchorscolors}}{
Configure the valuebox on the main page corresponding to the values for
failure, warning, success for fraction of unique query cells that
participate in anchor pairs. Failure corresponds to
[0:\code{Azimuth.map.fracanchorscolors[1]}), warning to
[\code{Azimuth.map.fracanchorscolors[1]}:\code{Azimuth.map.fracanchorscolors[2]}),
and success is >= \code{Azimuth.map.fracanchorscolors[2]}.
Defaults to \code{c(5, 15)}
}
\item{\code{Azimuth.map.postmapqccolors}}{
Configure the valuebox on the main page corresponding to the values for
failure, warning, success for the post mapping cluster based QC metric.
Failure corresponds to [0:\code{Azimuth.map.postmapqc[1]}), warning to
[\code{Azimuth.map.postmapqc[1]}:\code{Azimuth.map.postmapqc[2]}),
and success is >= \code{Azimuth.map.postmapqc[2]}.
Defaults to \code{c(0.15, 0.25)}
}
\item{\code{Azimuth.map.postmapqcds}}{
Set the amount of query random downsampling to perform before computing
the mapping QC metric.
Defaults to \code{5000}
}
\item{\code{Azimuth.map.ntrees}}{
Annoy (approximate nearest neighbor) n.trees parameter
Defaults to \code{20}
}
\item{\code{Azimuth.de.mincells}}{
Minimum number of cells per cluster for differential expression; defaults
to \code{15}
}
\item{\code{Azimuth.de.digits}}{
Number of digits to round differential expression table to; defaults to
\code{3}
}
\item{\code{Azimuth.sct.ncells}, \code{Azimuth.sct.nfeats}}{
Number of cells and features to use for
\code{\link[Seurat]{SCTransform}}, respectively. Defaults to \code{2000}
for each
}
}
}

\subsection{External options}{
The following options are used by external dependencies that have an effect
on \pkg{Azimuth}'s behavior. Refer to original package documentation for
more details
\describe{
\item{\code{\link[shiny:shiny-options]{shiny.maxRequestSize}}}{
User-configurable; used for controlling the maximum file size of uploaded
datasets. Defaults to 500 Mb
}
\item{\code{\link[DT:datatable]{DT.options}}}{
User-configurable; used for controlling biomarker table outputs.
Defaults to setting \code{pageLength} to \code{10}
}
\item{\code{\link[future:future.options]{future.globals.maxSize}}}{
\strong{Non-configurable}; used for parallelization. Defaults to
\code{Azimuth.app.max_cells * 320000}
}
}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/satijalab/azimuth}
  \item Report bugs at \url{https://github.com/satijalab/azimuth/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{phoffman@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

Authors:
\itemize{
  \item Andrew Butler \email{abutler@nygenome.org} (\href{https://orcid.org/0000-0003-3608-0463}{ORCID})
  \item Charlotte Darby \email{cdarby@nygenome.org} (\href{https://orcid.org/0000-0003-2195-5300}{ORCID})
  \item Yuhan Hao \email{yhao@nygenome.org} (\href{https://orcid.org/0000-0002-1810-0822}{ORCID})
  \item Rahul Satija \email{rsatija@nygenome.org} (\href{https://orcid.org/0000-0001-9448-8833}{ORCID})
}

Other contributors:
\itemize{
  \item Jaison Jain \email{jjain@nygenome.org} (\href{https://orcid.org/0000-0002-9478-5018}{ORCID}) [contributor]
  \item Satija Lab and Collaborators [funder]
}

}
