% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{AzimuthApp}
\alias{AzimuthApp}
\title{Launch the mapping app}
\usage{
AzimuthApp(config = NULL, ...)
}
\arguments{
\item{config}{Path to JSON-formatted configuration file specifying options.}

\item{...}{Options to set}
}
\value{
None, launches the mapping Shiny app
}
\description{
Launch the mapping app
}
\section{Specifying options}{

R options can be provided as named arguments to AzimuthApp through dots (...),
set in a config file, or set globally. Arguments provided to AzimuthApp
through dots take precedence if the same option is provided in a config file.
Options provided through dots or a config file take precedence if the same
option was set globally.
}

\seealso{
\code{\link{Azimuth-package}}
}
