% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{AzimuthBridgeServer}
\alias{AzimuthBridgeServer}
\title{Server function for the mapping app for bridge integration}
\usage{
AzimuthBridgeServer(input, output, session)
}
\arguments{
\item{input, output, session}{Required Shiny app server parameters}
}
\value{
The shiny server logic
}
\description{
Server function for the mapping app for bridge integration
}
\keyword{internal}
