% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_stuff.R
\docType{class}
\name{AzimuthData-class}
\alias{AzimuthData-class}
\alias{AzimuthData}
\title{The AzimuthData class is used to store reference info needed for Azimuth}
\description{
The AzimuthData class is used to store reference info needed for Azimuth
}
\section{Slots}{

\describe{
\item{\code{plotref}}{DimReduc object containing UMAP for plotting and projection.
This should also contain the cell IDs in the misc slot}

\item{\code{avgref}}{Average RNA expression for pseudobulk correlation tests}

\item{\code{colormap}}{Vector of id-color mapping for specifying the plots.}
}}

