% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{AzimuthToSeurat}
\alias{AzimuthToSeurat}
\title{Transfer downloaded Azimuth results into Seurat object (or create a new one)}
\usage{
AzimuthToSeurat(object = NULL, azimuth_results = NULL)
}
\arguments{
\item{object}{existing Seurat object to add azimuth_results to (creates new object if not given)}

\item{azimuth_results}{result of downloading "All results" from Azimuth,
or path to file containing these results}
}
\description{
Transfer downloaded Azimuth results into Seurat object (or create a new one)
}
