% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{CalcMappingMetric}
\alias{CalcMappingMetric}
\title{Calculate a mapping metric}
\usage{
CalcMappingMetric(object, reduction = "int", dims = 1:50)
}
\arguments{
\item{object}{A \code{Seurat} object with reference and query cells; easily
generated with \code{\link{QueryReference}}}

\item{reduction}{Name of reduction to use}

\item{dims}{Which dimensions to use}
}
\value{
Returns a data frame with the following columns:
\describe{
 \item{\code{mapping.score}}{Mapping score}
 \item{\code{mapping.score.smoothed}}{Smoothed mapping score}
 \item{\code{r.metric}}{R metric}
 \item{\code{mapped}}{
  A logical indiciating if the cell has been mapped or not
 }
 This data frame has the same number of cells marked as \dQuote{query} in
 \code{object$ingest}; rownames are the names of query cells
}
}
\description{
Calculate a mapping metric
}
\keyword{internal}
