% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{CategoryTable}
\alias{CategoryTable}
\title{Returns a dataframe of the frequency or percentage of levels of category.2
(column) for object split by each level of category.1 (row)}
\usage{
CategoryTable(object, category.1, category.2, percentage = FALSE)
}
\arguments{
\item{object}{a Seurat object}

\item{category.1}{a metadata field in the object}

\item{category.2}{another metadata field in the object}

\item{percentage}{if TRUE, returns percentages; otherwise, counts}
}
\description{
Returns a dataframe of the frequency or percentage of levels of category.2
(column) for object split by each level of category.1 (row)
}
\keyword{internal}
