% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ConvertGeneNames}
\alias{ConvertGeneNames}
\title{Converts feature names of query to match type/species of reference names}
\usage{
ConvertGeneNames(object, reference.names, linked)
}
\arguments{
\item{object}{object to convert, must contain only RNA counts matrix}

\item{reference.names}{rownames of reference}

\item{linked}{table with human/mouse homologies}
}
\value{
query object with converted feature names, likely subsetted
}
\description{
Converts feature names of query to match type/species of reference names
}
