% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{CreateAnn}
\alias{CreateAnn}
\title{Create an annoy index}
\usage{
CreateAnn(name, ndim)
}
\arguments{
\item{name, ndim}{Paramters}
}
\value{
An nn index object
}
\description{
Create an annoy index
}
\note{
Function exists because it's not exported from \pkg{uwot}
}
\keyword{internal}
