% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth.R
\name{CreateAzimuthData}
\alias{CreateAzimuthData}
\title{Create an \code{\link{AzimuthData}} object}
\usage{
CreateAzimuthData(
  object,
  plotref = "umap",
  plotids = NULL,
  avgref = NULL,
  colormap = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{plotref}{Either the name of the DimReduc in the provided Seurat object
to use for the plotting reference or the DimReduc object itself.}

\item{plotids}{Vector of IDs specifying the identities of the cells in the
plotref.}

\item{avgref}{Matrix containing the average RNA expression for the reference,
used in the pseudobulk correlation test.}

\item{colormap}{A named and ordered vector specifying the colors and levels
for the IDs being plotted. See \code{\link{CreateColorMap}} for help
generating your own.}
}
\value{
Returns an \code{\link{AzimuthData}} object
}
\description{
Create an auxiliary \code{\link{AzimuthData}} object for storing necessary
info when generating an Azimuth reference.
}
