% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth.R
\name{CreateColorMap}
\alias{CreateColorMap}
\title{Create A Color Map}
\usage{
CreateColorMap(object, ids = NULL, colors = NULL, seed = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{ids}{Vector of IDs to link to colors}

\item{colors}{Vector of colors to use}

\item{seed}{Set to randomly shuffle color assignments}
}
\value{
A named vector of colors
}
\description{
Create mapping between IDs and colors to use with reference plotting in
Azimuth
}
