% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.R
\name{dimPlotServer}
\alias{dimPlotServer}
\title{Create a \code{\link[Seurat]{DimPlot}} Shiny Module}
\usage{
dimPlotServer(id, object, ...)
}
\arguments{
\item{id}{...}

\item{object}{A \link[shiny]{reactive} \code{\link[Seurat]{Seurat}} object}
}
\value{
...
}
\description{
Create a \code{\link[Seurat]{DimPlot}} Shiny Module
}
