% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{dimPlotUI}
\alias{dimPlotUI}
\title{View a \code{\link[Seurat]{DimPlot}} in a Shiny App}
\usage{
dimPlotUI(id)
}
\arguments{
\item{id}{...}
}
\value{
...
}
\description{
View a \code{\link[Seurat]{DimPlot}} in a Shiny App
}
