% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{FilterFeatures}
\alias{FilterFeatures}
\title{Sanitize feature names for \code{\link[shiny]{selectInput}}}
\usage{
FilterFeatures(features)
}
\arguments{
\item{features}{A character vector of feature names}
}
\value{
\code{features}, but sanitized
}
\description{
\code{\link[shiny]{selectInput}} has some limitiations with biological
feature names. This function sanitizes feature names according to the
following rules:
\itemize{
 \item Names matching the regular expression \dQuote{\\.\\d+$} are
  \strong{removed}
}
}
\seealso{
\code{\link[shiny]{selectInput}}
}
\keyword{internal}
