% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{FormatDiffTime}
\alias{FormatDiffTime}
\title{Format Time Differences}
\usage{
FormatDiffTime(dt)
}
\arguments{
\item{dt}{A \code{\link[base]{difftime}} object}
}
\value{
The time difference in a nice string
}
\description{
Format Time Differences
}
\seealso{
\code{\link[base:difftime]{base::difftime}}
}
\keyword{internal}
