% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth.R
\name{GetAvgRef}
\alias{GetAvgRef}
\alias{GetAvgRef.AzimuthData}
\alias{GetAvgRef.Seurat}
\title{Get Azimuth average expression}
\usage{
GetAvgRef(object, ...)

\method{GetAvgRef}{AzimuthData}(object, ...)

\method{GetAvgRef}{Seurat}(object, slot = "AzimuthReference", ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{slot}{Name of tool}
}
\value{
A feature x 1 matrix of average RNA expression values
}
\description{
Get Azimuth average expression
}
