% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth.R
\name{GetColorMap}
\alias{GetColorMap}
\alias{GetColorMap.AzimuthData}
\alias{GetColorMap.Seurat}
\title{Get Azimuth color mapping}
\usage{
GetColorMap(object, ...)

\method{GetColorMap}{AzimuthData}(object, ...)

\method{GetColorMap}{Seurat}(object, slot = "AzimuthReference", ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{slot}{Name of tool}
}
\value{
A named vector specifying the colors for all reference IDs
}
\description{
Pull ID-color mapping for Azimuth plotting
}
