% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_stuff.R
\name{GetPlotRef}
\alias{GetPlotRef}
\alias{GetPlotRef.AzimuthData}
\alias{GetPlotRef.Seurat}
\title{Get Azimuth plotref}
\usage{
GetPlotRef(object, ...)

\method{GetPlotRef}{AzimuthData}(object)

\method{GetPlotRef}{Seurat}(object, slot = "AzimuthReference")
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{slot}{Name of tool}
}
\value{
A DimReduc object
}
\description{
Pull DimReduc used in Azimuth plotting/projection
}
