% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth.R
\name{GetSdRef}
\alias{GetSdRef}
\alias{GetSdRef.AzimuthData}
\alias{GetSdRef.Seurat}
\title{Get Azimuth average expression}
\usage{
GetSdRef(object, ...)

\method{GetSdRef}{AzimuthData}(object, ...)

\method{GetSdRef}{Seurat}(object, slot = "AzimuthReference", ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{slot}{Name of tool}
}
\value{
A feature x 1 matrix of average RNA expression values
}
\description{
Pull reference stdev RNA expression matrix
}
