% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{initQC}
\alias{initQC}
\title{Set up uploaded or demo file for QC}
\usage{
initQC(app.env, refs, session, output, googlesheet, mt.key, mito.pattern)
}
\arguments{
\item{app.env}{App environment}

\item{refs}{Reference object as from \code{\link{LoadReference}}}

\item{session}{Shiny app session}

\item{output}{Shiny app output}

\item{googlesheet}{Google sheet object as from \code{googlesheets4::gs4_get()}}

\item{mt.key}{String identifier for mitochondrial content}

\item{mito.pattern}{String pattern for identifying mitochondrial genes}
}
\description{
Set up uploaded or demo file for QC
}
