% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{LoadFileInput}
\alias{LoadFileInput}
\title{Load file input into a \code{Seurat} object}
\usage{
LoadFileInput(path)
}
\arguments{
\item{path}{Path to input data}
}
\value{
A \code{\link[Seurat]{Seurat}} object
}
\description{
Take a file and load it into a \code{\link[Seurat]{Seurat}} object. Supports
a variety of file types and always returns a \code{Seurat} object
}
\keyword{internal}
