% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{LoadH5ADobs}
\alias{LoadH5ADobs}
\title{Load obs from a H5AD file}
\usage{
LoadH5ADobs(path)
}
\description{
Read in only the metadata of an H5AD file and
return a data.frame object
}
\section{AnnData H5AD File (extension \code{h5ad})}{

}

