% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{LoadReference}
\alias{LoadReference}
\title{Load the reference RDS files}
\usage{
LoadReference(path, seconds = 10L)
}
\arguments{
\item{path}{Path or URL to the two RDS files}

\item{seconds}{Timeout to check for URLs in seconds}
}
\value{
A list with two entries:
\describe{
 \item{\code{reference}}{The reference \code{\link[Seurat]{Seurat}} object}
 \item{\code{index}}{
   A list with index information, includes an
   \code{\link[RcppAnnoy]{AnnoyIndex}} object
 }
}
}
\description{
Read in a reference \code{\link[Seurat]{Seurat}} object and annoy index. This
function can read either from URLs or a file path. In order to read properly,
there must be the following files:
\itemize{
 \item \dQuote{ref.Rds} for the reference \code{Seurat} object
 \item \dQuote{idx.Rds} for the R-based index object
 \item \dQuote{idx.annoy} for the
 \link{RcppAnnoy::AnnoyIndex}{RcppAnnoy-based} index object
}
}
\examples{
\dontrun{
# Load from a URL
ref <- LoadReference("http://saucyx220.nygenome.org")
# Load from a directory
ref2 <- LoadReference("/var/www/html")
}

}
\keyword{internal}
