% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{MappingScore}
\alias{MappingScore}
\title{Metric for evaluating transfer quality}
\usage{
MappingScore(
  anchors,
  combined.object,
  query.neighbors,
  ref.embeddings,
  query.embeddings,
  k = 50,
  ndim = 50,
  ksmooth = 100,
  ksnn = 20,
  snn.prune = 0,
  subtract.first.nn = TRUE,
  approx = FALSE,
  query.weights = NULL,
  verbose = TRUE
)
}
\arguments{
\item{combined.object}{Shell combined object (ref + query) from the
Anchorset object return}

\item{query.neighbors}{Neighbors object computed on query cells}

\item{ref.embeddings}{Reference embeddings matrix}

\item{query.embeddings}{Query embeddings matrix}

\item{k}{Number of anchors to use in projection steps when computing weights}

\item{ndim}{Number of dimensions to use when working with low dimensional
projections of the data}

\item{ksmooth}{Number of cells to average over when computing transition
probabilities}

\item{ksnn}{Number of cells to average over when determining the kernel
bandwidth from the SNN graph}

\item{snn.prune}{Amount of pruning to apply to edges in SNN graph}

\item{approx}{Use annoy rather than RANN in scoring}

\item{query.weights}{Query weights matrix for reuse}

\item{verbose}{Display messages/progress}

\item{anchorset}{Anchor matrix from the Anchorset object returned from
FindTransferAnchors}

\item{substract.first.nn}{Option to the scoring function when computing
distances to subtract the distance to the first nearest neighbor}
}
\value{
Returns a vector of cell scores
}
\description{
Metric for evaluating transfer quality
}
