% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{MinimalMatchingCells}
\alias{MinimalMatchingCells}
\title{Find a minimum number of matching cells}
\usage{
MinimalMatchingCells(reference, query, match = TRUE, seed = NULL)
}
\arguments{
\item{reference}{Reference \code{\link[Seurat]{Seurat}} object}

\item{query}{Query \code{\link[Seurat]{Seurat}} object}

\item{match}{Ensure that we only use identity classes present in both objects}

\item{seed}{a single value, interpreted as an integer, or \code{NULL}
    (see \sQuote{Details}).}
}
\value{
A vector of cells in \code{reference} in the identites of
\code{query}; for each identity, there will be either the same number of
cells for that identity as in \code{reference} or in \code{query}, whichever
is smaller
}
\description{
Find a minimum number of matching cells
}
\keyword{internal}
