% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{NNTransform}
\alias{NNTransform}
\title{Transform an NN index}
\usage{
NNTransform(neighbors, meta.data, key = "ori.index")
}
\arguments{
\item{neighbors}{Neighbors}

\item{meta.data}{Metadata}

\item{key}{Column of metadata to use}
}
\value{
\code{neighbors} transfomed
}
\description{
Transform an NN index
}
\keyword{internal}
