% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Oxford}
\alias{Oxford}
\title{Make An English List}
\usage{
Oxford(..., join = c("and", "or"))
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}

\item{join}{either \dQuote{and} or \dQuote{or}}
}
\value{
A character vector of the values, joined together with commas and
\code{join}
}
\description{
Joins items together to make an English list; uses the Oxford comma for the
last item in the list.
}
\examples{
\donttest{
Oxford("red")
Oxford("red", "blue")
Oxford("red", "green", "blue")
}

}
\keyword{internal}
