% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{PBCorTest}
\alias{PBCorTest}
\title{Pseudobulk correlation test of query against reference}
\usage{
PBCorTest(object, ref, min.features = 250)
}
\arguments{
\item{object}{Query object}

\item{ref}{Reference expression averages}

\item{min.features}{If fewer than min.features exist in the intersection,
return 0 as correlation.}
}
\value{
Returns correlation between query and reference
}
\description{
Computes the correlation between a pseudobulk of the query object and the
reference dataset. The feature set is the intersection of the reference
variable features and all features present in the query. Correlation is
computed on log normalized expression values using spearman correlation.
}
