% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{QueryReference}
\alias{QueryReference}
\title{Merge reference and query objects together}
\usage{
QueryReference(
  reference,
  query,
  assay.query = "RNA",
  reduction.reference = "spca",
  reduction.query = "int",
  reduction.name = "int",
  dims = 1:50,
  ...
)
}
\arguments{
\item{reference}{Reference \code{\link[Seurat]{Seurat}} object}

\item{query}{Query \code{\link[Seurat]{Seurat}} object}

\item{assay.query}{Name of assay in \code{query} to use}

\item{reduction.reference}{Name of reduction in \code{reference} to use}

\item{reduction.query}{Name of reduction in \code{query} to use}

\item{reduction.name}{Name to store dimensional reduction under in the Seurat object}

\item{dims}{Which dimensions to use}

\item{...}{
  Arguments passed on to \code{\link[=MinimalMatchingCells]{MinimalMatchingCells}}
  \describe{
    \item{\code{match}}{Ensure that we only use identity classes present in both objects}
    \item{\code{seed}}{a single value, interpreted as an integer, or \code{NULL}
    (see \sQuote{Details}).}
  }}
}
\value{
A \code{\link[Seurat]{Seurat}} object with the merged reference and
query datasets
}
\description{
Merge reference and query objects together
}
\keyword{internal}
