% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth.R
\name{RefDRDP}
\alias{RefDRDP}
\title{RefDR Dot Product}
\usage{
RefDRDP(query, ref, query.dr, ref.dr, grouping.var)
}
\arguments{
\item{query}{Query object}

\item{ref}{Reference object}

\item{query.dr}{Query dimensional reduction}

\item{ref.dr}{Reference dimensional reduction}

\item{grouping.var}{Grouping variable for the reference}
}
\description{
Mapping metric based on dot products of projected PCAs
}
