% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth.R
\name{ReferenceVersion}
\alias{ReferenceVersion}
\alias{ReferenceVersion.AzimuthData}
\alias{ReferenceVersion.Seurat}
\title{Get Azimuth reference version number}
\usage{
ReferenceVersion(object, ...)

\method{ReferenceVersion}{AzimuthData}(object, ...)

\method{ReferenceVersion}{Seurat}(object, slot = "AzimuthReference", ...)
}
\arguments{
\item{object}{Seurat or AzimuthData object}

\item{...}{Not used}

\item{slot}{Name of the version to pull. Can be "seurat.version",
"azimuth.version", or "reference.version".}
}
\value{
A character string specifying the reference version
}
\description{
Pull the reference version information
}
