% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{RenderDiffExp}
\alias{RenderDiffExp}
\title{Prepare differential expression results for rendering}
\usage{
RenderDiffExp(diff.exp, groups.use = NULL, n = 10L, logfc.thresh = 0L)
}
\arguments{
\item{diff.exp}{A dataframe with differential expression results from
\code{\link{presto::wilcoxauc}}}

\item{groups.use}{Names of groups to filter \code{diff.exp} to; groups must
be found in \code{diff.exp$group}}

\item{n}{Number of feature to filter \code{diff.exp} to per group}

\item{logfc.thresh}{logFC threshold}
}
\value{
\code{diff.exp}, ordered by adjusted p-value, filtered to \code{n}
features per group in \code{group.use}
}
\description{
Prepare differential expression results for rendering
}
\seealso{
\code{\link[presto]{wilcoxauc}}
}
\keyword{internal}
