% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{RenderDiffMotifExp}
\alias{RenderDiffMotifExp}
\title{Prepare differential expression motif results for rendering}
\usage{
RenderDiffMotifExp(diff.exp, groups.use = NULL, n = 10L)
}
\arguments{
\item{diff.exp}{A dataframe with differential expression results from 
\code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers}}}

\item{groups.use}{Names of groups to filter \code{diff.exp} to; groups must
be found in \code{diff.exp$cluster}}

\item{n}{Number of feature to filter \code{diff.exp} to per group}
}
\value{
\code{diff.exp}, ordered by adjusted p-value, filtered to \code{n}
features per group in \code{group.use}
}
\description{
Prepare differential expression motif results for rendering
}
\seealso{
\code{\link[Seurat]{FindAllMarkers}}
}
\keyword{internal}
