% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth.R, R/generics.R
\name{RunAzimuth.Seurat}
\alias{RunAzimuth.Seurat}
\alias{RunAzimuth.character}
\alias{RunAzimuth}
\title{Run Azimuth annotation}
\usage{
\method{RunAzimuth}{Seurat}(
  query,
  reference,
  annotation.levels = NULL,
  umap.name = "ref.umap",
  do.adt = FALSE,
  verbose = TRUE,
  assay = "RNA",
  k.weight = 50,
  n.trees = 20,
  mapping.score.k = 100
)

\method{RunAzimuth}{character}(query, ...)

RunAzimuth(query, ...)
}
\arguments{
\item{query}{Seurat object or following type of path:
\itemize{
 \item A \code{.h5} matrix
 \item A \code{.rds} file containing a Seurat object
 \item A \code{.h5ad} anndata object
 \item A \code{.h5seurat} object
}}

\item{reference}{Name of reference to map to or a path to a directory containing ref.Rds and idx.annoy}

\item{annotation.levels}{list of annotation levels to map. If not specified, all will be mapped.}

\item{umap.name}{name of umap reduction in the returned object}

\item{do.adt}{transfer ADT assay}

\item{assay}{query assay name}
}
\value{
Seurat object with reference reductions and annotations

Returns a Seurat object containing celltype annotations
}
\description{
Run Azimuth annotation
}
