% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth.R, R/generics.R
\name{RunAzimuthATAC.Seurat}
\alias{RunAzimuthATAC.Seurat}
\alias{RunAzimuthATAC.character}
\alias{RunAzimuthATAC}
\title{Run Azimuth annotation for ATAC query}
\usage{
\method{RunAzimuthATAC}{Seurat}(
  query,
  reference,
  fragment.path = NULL,
  annotation.levels = NULL,
  umap.name = "ref.umap",
  verbose = TRUE,
  assay = NULL,
  k.weight = 50,
  n.trees = 20,
  mapping.score.k = 100,
  dims.atac = 2:50,
  dims.rna = 1:50
)

\method{RunAzimuthATAC}{character}(query, ...)

RunAzimuthATAC(query, ...)
}
\arguments{
\item{query}{Seurat object or following type of path:
\itemize{
 \item A \code{.h5} matrix
 \item A \code{.rds} file containing a Seurat object
 \item A \code{.h5ad} anndata object
 \item A \code{.h5seurat} object
}}

\item{reference}{Name of reference to map to or a path to a directory containing ext.Rds}

\item{annotation.levels}{list of annotation levels to map. If not specified, all will be mapped.}

\item{umap.name}{name of umap reduction in the returned object}

\item{assay}{query assay name}

\item{dims.atac}{dimensions}

\item{dims.rna}{dimensions}

\item{do.adt}{transfer ADT assay}
}
\value{
Seurat object with reference reductions and annotations

Returns a Seurat object containing celltype annotations
}
\description{
Run Azimuth annotation for ATAC query
}
