% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{SaveAzimuthReference}
\alias{SaveAzimuthReference}
\title{Save \code{Azimuth} references and neighbors index to same folder}
\usage{
SaveAzimuthReference(object = NULL, folder = NULL)
}
\arguments{
\item{object}{An \code{\link{Azimuth}} reference}

\item{file}{Path to save \code{Azimuth} reference to; defaults to
\code{file.path(getwd(), "azimuth_reference/"))}}

\item{...}{
  Arguments passed on to \code{\link[base:readRDS]{base::saveRDS}}
  \describe{
    \item{\code{ascii}}{a logical.  If \code{TRUE} or \code{NA}, an ASCII
    representation is written; otherwise (default), a binary one is used.
    See the comments in the help for \code{\link[base]{save}}.}
    \item{\code{version}}{the workspace format version to use.  \code{NULL}
    specifies the current default version (3).  The only other supported
    value is 2, the default from \R 1.4.0 to \R 3.5.0.}
    \item{\code{compress}}{a logical specifying whether saving to a named file is
    to use \code{"gzip"} compression, or one of \code{"gzip"},
    \code{"bzip2"}, \code{"xz"} or \code{"zstd"} to indicate the type of
    compression to be used.  Ignored if \code{file} is a connection.}
    \item{\code{refhook}}{a hook function for handling reference objects.}
  }}
}
\value{
Invisibly returns \code{file}
}
\description{
Save \code{Azimuth} references and neighbors index to same folder
}
\examples{
# Make Azimuth Reference object
obj.azimuth <- AzimuthReference(object)

# Save 
SaveAzimuthReference(object = obj.azimuth, folder = "azimuth_reference")

# Run Azimuth

query <- RunAzimuth(query = query, 
                    reference = "azimuth_reference", 
                    ...)


}
\seealso{
\code{\link{saveRDS}()} \code{\link{readRDS}()}
}
