% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth.R
\name{SetColorMap}
\alias{SetColorMap}
\alias{SetColorMap.AzimuthData}
\alias{SetColorMap.Seurat}
\title{Set Azimuth color mapping}
\usage{
SetColorMap(object, ...)

\method{SetColorMap}{AzimuthData}(object, value, ...)

\method{SetColorMap}{Seurat}(object, slot = "AzimuthReference", value, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{value}{New colormap to assign}

\item{slot}{Name of tool}
}
\value{
An object with the colormap slot set
}
\description{
Set ID-color mapping for Azimuth plotting
}
