% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SeuratMapper-package}
\alias{SeuratMapper-package}
\alias{_PACKAGE}
\alias{SeuratMapper}
\title{SeuratMapper: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\section{Package options}{


SeuratMapper uses the following options to control the behaviour of the app,
users can configure these with \code{\link[base]{options}}:

\describe{
 \item{\code{Azimuth.de.mincells}}{
  Minimum number of cells per cluster for differential expression; defaults
  to \code{15}
 }
 \item{\code{Azimuth.map.pcthresh}}{
  Only show mapped plot if the percentage of cells mapped meets or
  exceeds this threshold; defaults to \code{60}
 }
 \item{\code{Azimuth.sct.ncells}, \code{Azimuth.sct.nfeats}}{
  Number of cells and features to use for
  \code{\link[Seurat]{SCTransform}}, respectively. Defaults to \code{1000}
  for each
 }
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mojaveazure/seurat-mapper}
  \item Report bugs at \url{https://github.com/mojaveazure/seurat-mapper/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{phoffman@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

Authors:
\itemize{
  \item Yuhan Hao \email{yhao@nygenome.org} (\href{https://orcid.org/0000-0002-1810-0822}{ORCID})
  \item Rahul Satija \email{rsatija@nygenome.org} (\href{https://orcid.org/0000-0001-9448-8833}{ORCID})
  \item Charlotte Darby \email{cdarby@nygenome.org} (\href{https://orcid.org/0000-0003-2195-5300}{ORCID})
}

}
