% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{TabJS}
\alias{TabJS}
\alias{TabJSHide}
\alias{TabJSKey}
\title{Manage tabs with \pkg{shinyjs}}
\usage{
TabJSHide(id, values, fxn = "hide")

TabJSKey(id, values)
}
\arguments{
\item{id}{ID of a \code{\link[shiny]{tabsetPanel}}}

\item{values}{One or more values of a \code{\link[shiny]{tabPanel}}
(see the \code{value} parameter)}

\item{fxn}{Name of JavaScript call function}
}
\value{
\code{TabJSHide}: a string with a JavaScript function to hide a set
of tabs

\code{TabJSKey}: a string with the JavaScript ID for a given
set of tabs
}
\description{
Quickly generate JavaScript IDs for Shiny tab panels. Also build a function
to hide Shiny tab panels in JavaScript using \pkg{shinyjs}
}
\note{
These functions are designed to run custom JavaScript code using
\code{\link[shinyjs]{extendShinyJS}}; use of custom JavaScript code requires
the \pkg{V8} package. \pkg{V8} requires a local install of either the
\href{https://chromium.googlesource.com/v8/v8}{V8} JavaScript Engine or
\href{https://nodejs.org/en/}{Node.js}
}
\examples{
\dontrun{
TabJSHide('tabs', values = c('mapped', 'fexplorer'))
}

\dontrun{
TabJSKey('tabs', values = 'mapped')
}

}
\seealso{
\code{\link[shinyjs]{extendShinyJS}} \code{\link[shiny]{tabPanel}}
}
\keyword{internal}
