% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.R
\name{tenxFileServer}
\alias{tenxFileServer}
\alias{tenxH5Server}
\title{Load a 10X Dataset}
\usage{
tenxH5Server(id, ...)
}
\arguments{
\item{id}{...}
}
\value{
A \link[shiny]{reactive} \code{\link[Seurat]{Seurat}} object
}
\description{
Load a 10X Dataset
}
