% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_stuff.R
\name{ValidateAzimuthReference}
\alias{ValidateAzimuthReference}
\title{Validate aspects of a Seurat object to be used as an Azimuth reference}
\usage{
ValidateAzimuthReference(object, ad.name = "AzimuthReference")
}
\arguments{
\item{object}{Seurat object}

\item{ad.name}{Name in the tools slot containing the AzimuthData object.}
}
\value{
No return value
}
\description{
Validate aspects of a Seurat object to be used as an Azimuth reference
}
