% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{AddPredictions}
\alias{AddPredictions}
\title{Add Predictions}
\usage{
AddPredictions(
  object,
  preds,
  scores,
  preds.levels = unique(x = as.character(x = preds)),
  preds.drop = TRUE
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{preds}{A named vector with the predicted IDs}

\item{scores}{A named vector with the scores for \code{preds}}

\item{preds.levels}{Levels for predicted IDs, useful for
ordering \code{preds}}

\item{preds.drop}{Drop unused levels from \code{preds}}
}
\value{
\code{object} with transformations applied
}
\description{
Add Predictions
}
\keyword{internal}
