% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{AzimuthApp}
\alias{AzimuthApp}
\title{Launch the mapping app}
\usage{
AzimuthApp(reference = "http://satijalab04.nygenome.org/pbmc")
}
\arguments{
\item{reference}{Path or URL to reference dataset}
}
\value{
None, launches the mapping Shiny app
}
\description{
Launch the mapping app
}
