% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth.R
\name{AzimuthApp}
\alias{AzimuthApp}
\title{Launch the mapping app}
\usage{
AzimuthApp(config = NULL, type = "standard", ...)
}
\arguments{
\item{config}{Path to JSON-formatted configuration file specifying options;
for an example config file, see
\code{system.file("resources", "config.json", package = "Azimuth")}}

\item{...}{Options to set, see \code{?`\link{Azimuth-package}`} for details
on \pkg{Azimuth}-provided options}
}
\value{
None, launches the mapping Shiny app
}
\description{
Launch the mapping app
}
\section{Specifying options}{

R options can be provided as named arguments to \code{AzimuthApp} through
dots (...), set in a config file, or set globally. Arguments provided to
\code{AzimuthApp} through dots take precedence if the same option is provided
in a config file. Options provided through dots or a config file take
precedence if the same option was set globally.

Options in the \code{\link[Azimuth:Azimuth-package]{Azimuth.app}} namespace
can be specified using a shorthand notation in both the config file and as
arguments to \code{AzimuthApp}. For example, the option
\code{Azimuth.app.reference} can be shortened to \code{reference} in the
config file or as an argument to \code{AzimuthApp}
}

\examples{
if (interactive()) {
  AzimuthApp(system.file("resources", "config.json", package = "Azimuth"))
}

}
\seealso{
\code{\link{Azimuth-package}}
}
