% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{AzimuthServer}
\alias{AzimuthServer}
\title{Server function for the mapping app}
\usage{
AzimuthServer(input, output, session)
}
\arguments{
\item{input, output, session}{Required Shiny app server parameters}
}
\value{
The shiny server logic
}
\description{
Server function for the mapping app
}
\keyword{internal}
