% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_stuff.R
\name{CreateAzimuthData}
\alias{CreateAzimuthData}
\title{Create auxiliary AzimuthData object for storing necessary info when generating
an Azimuth reference.}
\usage{
CreateAzimuthData(
  object,
  plotref = "umap",
  plotids = NULL,
  avgref = NULL,
  colormap = NULL,
  ad.name = "Azimuth"
)
}
\arguments{
\item{object}{Seurat object}

\item{plotref}{Either the name of the DimReduc in the provided Seurat object
to use for the plotting reference or the DimReduc object itself.}

\item{plotids}{Vector of IDs specifying the identities of the cells in the
plotref.}

\item{avgref}{Matrix containing the average RNA expression for the reference,
used in the pseudobulk correlation test.}

\item{colormap}{A named and ordered vector specifying the colors and levels
for the IDs being plotted. See \link{\code{CreateColorMap}} for help
generating your own.}

\item{ad.name}{Name in the tools slot to store the AzimuthData object.
Azimuth is expecting the name "Azimuth", so generally don't modify this.}
}
\value{
Returns a Seurat object with the AzimuthData object stored in the
tools slot
}
\description{
Create auxiliary AzimuthData object for storing necessary info when generating
an Azimuth reference.
}
