% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_stuff.R
\name{CreateColorMap}
\alias{CreateColorMap}
\title{Create mapping between IDs and colors to use with reference plotting in
Azimuth}
\usage{
CreateColorMap(object, ids = NULL, colors = NULL, seed = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{ids}{Vector of IDs to link to colors}

\item{colors}{Vector of colors to use}

\item{seed}{Set to randomly shuffle color assignments}
}
\description{
Create mapping between IDs and colors to use with reference plotting in
Azimuth
}
