% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{LoadBridgeReference}
\alias{LoadBridgeReference}
\title{Load the bridge, reference, and reference extension RDS files}
\usage{
LoadBridgeReference(path, seconds = 10L)
}
\arguments{
\item{path}{Path or URL to the RDS file}

\item{seconds}{Timeout to check for URLs in seconds}
}
\value{
A list with two entries:
\describe{
 \item{\code{map}}{
  The extended reference \code{\link[Seurat]{Seurat}}
  object 
 }
 \item{\code{plot}}{The reference \code{Seurat} object (for plotting)}
}
}
\description{
Read in a precomputed extended reference. This function can
read either from URLs or a file path. The function looks for a file 
called ext.Rds for the extended reference \code{Seurat} object
}
\examples{
\dontrun{
# Load from a URL
ref <- LoadBridgeReference("https://seurat.nygenome.org/references/pbmc")
# Load a file from the path to a directory 
ref2 <- LoadBridgeReference("path/")
# Load a file directly
ref3 <- LoadBridgeReference("ext.Rds")
}
}
