% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{LoadFileInput}
\alias{LoadFileInput}
\title{Load file input into a \code{Seurat} object}
\usage{
LoadFileInput(path)
}
\arguments{
\item{path}{Path to input data}
}
\value{
A \code{\link[Seurat]{Seurat}} object
}
\description{
Take a file and load it into a \code{\link[Seurat]{Seurat}} object. Supports
a variety of file types and always returns a \code{Seurat} object
}
\details{
\code{LoadFileInput} supports several file types to be read in as
\code{Seurat} objects. File type is determined by extension, matched in a
case-insensitive manner See sections below for details about supported
filtypes, required extension, and specifics for how data is loaded
}
\section{10X H5 File (extension \code{h5})}{

10X HDF5 files are supported for all versions of Cell Ranger; data is read
in using \code{\link[Seurat]{Read10X_h5}}. \strong{Note}: for multi-modal
10X HDF5 files, only the \emph{first} matrix is read in
}

\section{Rds File (extension \code{rds})}{

Rds files are supported as long as they contain one of the following data
types:
\itemize{
 \item A \code{\link[Seurat]{Seurat}} V3 object
 \item An S4 \code{\link[Matrix]{Matrix}} object
 \item An S3 \code{\link[base]{matrix}} object
 \item A \code{\link[base]{data.frame}} object
}
For S4 \code{Matrix}, S3 \code{matrix}, and \code{data.frame} objects, a
\code{Seurat} object will be made with
\code{\link[Seurat]{CreateSeuratObject}} using the default arguments
}

\section{h5Seurat File (extension \code{h5seurat})}{

h5Seurat files and all of their features are fully supported. They are read
in via \code{\link[SeuratDisk]{LoadH5Seurat}}. \strong{Note}: only the
\dQuote{counts} matrices are read in and only the default assay is kept
}

\section{AnnData H5AD File (extension \code{h5ad})}{

Only H5AD files from AnnData v0.7 or higher are supported. Data is read from
the H5AD file in the following manner
\itemize{
 \item The counts matrix is read from \dQuote{/raw/X}; if \dQuote{/raw/X} is
 not present, the matrix is read from \dQuote{/X}
 \item Feature names are read from feature-level metadata. Feature level
 metadata must be an HDF5 group, HDF5 compound datasets are \strong{not}
 supported. If counts are read from \code{/raw/X}, features names are looked
 for in \dQuote{/raw/var}; if counts are read from \dQuote{/X}, features
 names are looked for in \dQuote{/var}. In both cases, feature names are read
 from the dataset specified by the \dQuote{_index} attribute, \dQuote{_index}
 dataset, or \dQuote{index} dataset, in that order
 \item Cell names are read from cell-level metadata. Cell-level metadata must
 be an HDF5 group, HDF5 compound datasets are \strong{not} supported.
 Cell-level metadata is read from \dQuote{/obs}. Cell names are read from the
 dataset specified by the \dQuote{_index} attribute, \dQuote{_index} dataset,
 or \dQuote{index} dataset, in that order
 \item Cell-level metadata is read from the \dQuote{/obs} dataset. Columns
 will be returned in the same order as in the \dQuote{column-order}, if
 present, or in alphabetical order. If a dataset named \dQuote{__categories}
 is present, then all datasets in \dQuote{__categories} will serve as factor
 levels for datasets present in \dQuote{/obs} with the same name (eg. a
 dataset named \dQuote{/obs/__categories/leiden} will serve as the levels for
 \dQuote{/obs/leiden}). Row names will be set as cell names as described
 above. All datasets in \dQuote{/obs} will be loaded except for
 \dQuote{__categories} and the cell names dataset
}
}

