% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{LoadReference}
\alias{LoadReference}
\title{Load the reference RDS files}
\usage{
LoadReference(path, seconds = 10L)
}
\arguments{
\item{path}{Path or URL to the two RDS files}

\item{seconds}{Timeout to check for URLs in seconds}
}
\value{
A list with three entries:
\describe{
 \item{\code{map}}{
  The downsampled reference \code{\link[Seurat]{Seurat}}
  object (for mapping)
 }
 \item{\code{plot}}{The reference \code{Seurat} object (for plotting)}
 \item{\code{avgexp}}{Average expression (for pseudobulk check)}
}
}
\description{
Read in a reference \code{\link[Seurat]{Seurat}} object and annoy index. This
function can read either from URLs or a file path. In order to read properly,
there must be the following files:
\itemize{
 \item \dQuote{ref.Rds} for the downsampled reference \code{Seurat}
 object (for mapping)
 \item \dQuote{idx.annoy} for the nearest-neighbor index object
}
}
\examples{
\dontrun{
# Load from a URL
ref <- LoadReference("https://seurat.nygenome.org/references/pbmc")
# Load from a directory
ref2 <- LoadReference("/var/www/html")
}

}
\keyword{internal}
