% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{MappingScore}
\alias{MappingScore}
\title{Metric for evaluating transfer quality}
\usage{
MappingScore(
  anchorset,
  ref,
  query,
  ref.reduction = "spca",
  query.reduction = "spca",
  k = 50,
  ndim = 50,
  ksmooth = 100,
  ksnn = 20,
  snn.prune = 0,
  subtract.first.nn = TRUE,
  approx = FALSE,
  verbose = TRUE,
  debug = FALSE
)
}
\arguments{
\item{anchorset}{Anchorset object returned from FindTransferAnchors}

\item{ref}{Reference object}

\item{query}{Query object}

\item{k}{Number of anchors to use in projection steps when computing weights}

\item{ndim}{Number of dimensions to use when working with low dimensional
projections of the data}

\item{ksmooth}{Number of cells to average over when computing transition
probabilities}

\item{ksnn}{Number of cells to average over when determining the kernel
bandwidth from the SNN graph}

\item{snn.prune}{Amount of pruning to apply to edges in SNN graph}

\item{verbose}{Display messages/progress}
}
\value{
Returns a vector of cell scores
}
\description{
Metric for evaluating transfer quality
}
