% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{NNTransform}
\alias{NNTransform}
\title{Transform an NN index}
\usage{
NNTransform(
  object,
  meta.data,
  neighbor.slot = "query_ref.nn",
  key = "ori.index"
)
}
\arguments{
\item{object}{Seurat object}

\item{meta.data}{Metadata}

\item{neighbor.slot}{Name of Neighbor slot}

\item{key}{Column of metadata to use}
}
\value{
\code{object} with transfomed neighbor.slot
}
\description{
Transform an NN index
}
\keyword{internal}
